refactor(widget): shared partial-JSON parser + extractWidgetCode
